Crate bls_ringct

Re-exports

pub use bls_bulletproofs;

pub use bls_bulletproofs::blstrs;

pub use bls_bulletproofs::group;

pub use bls_bulletproofs::rand;

pub use error::Error;

pub use mlsag::DecoyInput;

pub use mlsag::MlsagMaterial;

pub use mlsag::MlsagSignature;

pub use mlsag::TrueInput;

pub use ringct::Output;

pub use ringct::RingCtMaterial;

Modules

error

mlsag

ringct

Structs

RevealedCommitment

Functions

hash_to_curve

Hashes a point to another point on the G1 curve

key_image

returns KeyImage for the given public/secret key pair A key image is defined to be I = x * Hp(P)

public_key

Type Definitions

Result